# src/main/resources/static/swagger.yaml
openapi: 3.0.3
info:
  title: Entrei
  description: |-
    API para controle de convidados
  version: 1.0.0
servers:
  - url: http://localhost
tags:
  - name: Evento
    description: Controle de eventos
  - name: QR Code
    description: Controle de QR codes
paths:
  /eventos:
    post:
      tags:
        - Evento
      summary: Cria um novo evento.
      operationId: postEvento
      requestBody:
        description: Dados para geração do evento.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventoModel'
        required: true
      responses:
        '200':
          description: Evento criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventoModel'
    get:
      tags:
        - Evento
      summary: Busca pelos eventos.
      operationId: getEventos
      responses:
        '200':
          description: Eventos encontrados com sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventoModel'

  /eventos/{id}:
    get:
      tags:
        - Evento
      summary: Busca evento pelo ID.
      operationId: getEvento
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Id do evento
      responses:
        '200':
          description: Evento encontrado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventoModel'
    delete:
      tags:
        - Evento
      summary: Exclui evento pelo ID.
      operationId: deleteEvento
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Id do evento
      responses:
        '200':
          description: Evento excluído com sucesso.

  /qr-codes:
    post:
      tags:
        - QR Code
      summary: Cria um novo QR code para o convidado.
      operationId: postQRCode
      requestBody:
        description: Dados para geração QR code.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QrCodeModel'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QrCodeModel'

components:
  schemas:
    EventoModel:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
          example: 'Casamento Gabriela e Pedro'
        descricao:
          type: string
          example: 'Celebração e festa de casamento'
        dataHora:
          type: string
          format: date-time
        responsaveis:
          type: array
          items:
            $ref: '#/components/schemas/PessoaModel'

    PessoaModel:
      type: object
      properties:
        nome:
          type: string
          example: 'Convidado X'
        email:
          type: string
          example: 'convidadox@mail.com'
        telefone:
          type: string
          example: '31977777777'

    QrCodeModel:
      type: object
      properties:
        id:
          type: string
        idEvento:
          type: string
        convidado:
          $ref: '#/components/schemas/PessoaModel'